split unwinding logic from panic logic
